# Perf: adds a new implementation for full joining multiple tables