This should allevieate performance issues with many variables as recently reported. In one test the time to do the full join went down from 4 seconds to 600 ms and the memory use from 250MB to 85MB

Slack discussion: https://owid.slack.com/archives/CQQUA2C2U/p1659005364871929

> The original motivation was performance improvements. Those were quite straightforward but then I kept running into breaking charts with the svg tester that led me to add several iterations of fixes (felt so close every time ). The end result is several improvements over the status quo and one open question:
> 
> **Benefits**
> 
> -   Performance - the chart initialization time for some heavy multi-variable charts improved from ~4000 ms join cpu time to ~600ms and a memory footprint reduction by about half. (e.g. this chart <https://ourworldindata.org/grapher/share-of-deaths-by-cause> that Pablo mentioned recently and Fiona complained about perf for similar charts)
> -   Day and Year variables are now joined properly. This means that you can now do scatter and marimekko charts that use day variables (e.g. covid vaccinations) and year variables (e.g. GDP/Population) and days are matched to year values of the corresponding year. One caveat: the year variable has to have values for the years the days are in (i.e. you need values for population or gdp for 2022 for this to show up correctly). For population we have a shared historical + predicted population variable that has values for 2022 but for GDP we might need such a unified variable for day based joins to work correctly. The problem when the year variable does not have values for the correct year is that the join picks an arbitrary year (usually the first in the time series). Properly fixing this would require some bigger refactorings. For now this means that the charts that use population with day variables should switch to the population version that has projections.
> -   Multi-variable joins are now more correct in some edge cases. E.g. population used to be joined after some other joins and for scatter plots could then often implicitly use the year for population that was picked in the join before. Now all years are lined up at once and special scatter logic only applies for variables that have a target year set. This also recovers a few countries that were previously dropped as they fell out of the tolerance ranges because of the old join logic - e.g. In <https://owid.cloud/admin/charts/867/edit> Yemen and Afghanistan have high values but are missing in the old join code; or Nigeria is missing currently in <https://owid.cloud/admin/charts/3041/edit> )
> 
> **Unwanted changes**
> 
> -   Continent colors are shifted in scatter plots that join day and year variables (i.e. covid scatter plots). This has to do with how Antarctica is treated that doesn't appear in the daily data series and is discarded in the new algorithm but was still present in the inputTable in the old algorithm. I am inclined to ignore this for now, especially as Marwa and Matt are working on improving color anyhow and then we'll probably want to implement stable mappings for continents anyhow. Thoughts?
> 
> You can see all differences in charts in the svg tester images for the branch: <https://github.com/owid/owid-grapher-svgs/commit/292fa892f0791bc8335b395adb8b717ca05d6512> (use swipe to easily see a larger version of each chart in the before/after view). For a few charts I will do some minor fixes after this is merged so that e.g. population will be used correctly for day variables.