@danyx23 2022-07-28
Charts to edit the config for when this is merged into prod: 
* https://owid.cloud/admin/charts/335/edit
* https://owid.cloud/admin/charts/5134/edit
* https://owid.cloud/admin/charts/379/edit