@danielgavrilov 2022-08-03
> Charts to edit the config for when this is merged into prod:
> 
> * https://owid.cloud/admin/charts/335/edit
> * https://owid.cloud/admin/charts/5134/edit
> * https://owid.cloud/admin/charts/379/edit

What needs to be edited for [335](https://owid.cloud/admin/charts/335/edit) to to work correctly? The before/after for context:

<img width="863" alt="Screenshot 2022-08-03 at 10 03 41" src="https://user-images.githubusercontent.com/1308115/182569611-e3571ac7-ac29-4499-9673-0e509421e554.png">