@marcelgerber 2022-08-03
> The problem when the year variable does not have values for the correct year is that the join picks an arbitrary year (usually the first in the time series).

This sounds a bit problematic maybe? It is certainly something that we need to communicate well. And as you say, it's problematic for GDP for example, where we probably don't even have projection data available.
Also, it's probably pretty common for a daily variable to have data well into the current year, and for a yearly variable to lag one or two years behind the current year.